+++ src/Models/Player.hpp
@@ -13,20 +13,20 @@
 class Player : public GameObject
 {
 public:
-	Player() : health(0), maxHealth(0), map(nullptr), velocity(0.0F)
+	Player() : health(0), maxHealth(0), map(nullptr), velocity(0)
 	{
 		this->objType = ObjectType::PLAYER;
 
-		this->lastMousePos = Vector2(0.0F, 0.0F);
+		this->lastMousePos = Vector2(0, 0);
 	}
 
-	Player(const Vector2<float> pos, const std::vector<Sprite> &sprites, const Vector2<float> size = Vector2<float>(Game_Parameters::SCREEN_HEIGHT / 15, Game_Parameters::SCREEN_HEIGHT / 15), int maxHealth = 100) : GameObject(pos, sprites[0], size, (int)ObjectType::PLAYER), map(nullptr), velocity(0.0F)
+	Player(const Vector2<int> pos, const std::vector<Sprite> &sprites, const Vector2<int> size = Vector2<int>(Game_Parameters::SCREEN_HEIGHT / 15, Game_Parameters::SCREEN_HEIGHT / 15), int maxHealth = 100) : GameObject(pos, sprites[0], size, (int)ObjectType::PLAYER), map(nullptr), velocity(0.0F)
 	{
 		this->maxHealth = maxHealth;
 		this->health = maxHealth;
 		this->sprites = sprites;
 
-		this->lastMousePos = Vector2(0.0F, 0.0F);
+		this->lastMousePos = Vector2(0, 0);
 	}
 
 	~Player();
@@ -38,24 +38,24 @@ public:
 	void ProcessInput();
 	void SetMap(Map *map);
 	void setSelectFalse();
-	void setVelocity(const float velocity);
+	void setVelocity(const int velocity);
 	void takeDamage(const int value);
 
 	void OnDestroy() override;
 	void Destroy() override;
 
-	float getVelocity() const;
+	int getVelocity() const;
 	int getHealth() const;
 	int getMaxHealth() const;
 
 private:
 	int health;
 	int maxHealth;
-	float velocity;
+	int velocity;
 
 	Map *map;
 
-	Vector2<float> lastMousePos;
+	Vector2<int> lastMousePos;
 
 	int currentIndex = 0;
 	std::vector<Sprite> sprites;
