+++ src/Models/Tile.cpp
@@ -4,11 +4,11 @@ Tile::Tile() : type(TileTypes::FLOOR), frame(-1), GameObject()
 {
 }
 
-Tile::Tile(Vector2<float> pos, const Sprite &sprite, Vector2<float> size, TileTypes type, int frame) : GameObject(pos, sprite, size, (int)ObjectType::TILE), type(type), frame(frame)
+Tile::Tile(Vector2<int> pos, const Sprite &sprite, Vector2<int> size, TileTypes type, int frame) : GameObject(pos, sprite, size, (int)ObjectType::TILE), type(type), frame(frame)
 {
 }
 
-Tile::Tile(Vector2<float> pos, const Sprite &sprite, Vector2<float> size) : GameObject(pos, sprite, size, (int)ObjectType::TILE), type(TileTypes::FLOOR), frame(-1)
+Tile::Tile(Vector2<int> pos, const Sprite &sprite, Vector2<int> size) : GameObject(pos, sprite, size, (int)ObjectType::TILE), type(TileTypes::FLOOR), frame(-1)
 {
 }
